feat: configed routes for landing page
