+++ src/pages/Dashboard/Sidebar.tsx
@@ -55,7 +55,7 @@ const Sidebar: React.FC<SidebarProps> = ({ sidebarItemGroups, sidebarheader }) =
         <SidebarHeader>
           <SidebarMenu>
             <SidebarMenuItem>
-              <SidebarMenuButton asChild className="h-auto bg-accent border hover:bg-secondary-light rounded-full" size={"lg"}>
+              <SidebarMenuButton asChild className="h-auto bg-background border hover:bg-secondary-light rounded-full" size={"lg"}>
                 <Link to={sidebarheader.url ?? ""}>
                   <div className="w-full flex items-center gap-2 justify-between">
                     <div className="flex items-center gap-2">
